feat: delete player needs to delete decks...

+++ src/main/java/org/springframework/samples/petclinic/player/PlayerController.java
@@ -15,7 +15,6 @@ import org.springframework.samples.petclinic.exceptions.ResourceNotFoundExceptio
 import org.springframework.samples.petclinic.user.User;
 import org.springframework.samples.petclinic.user.UserService;
 import org.springframework.samples.petclinic.util.RestPreconditions;
-import org.springframework.samples.petclinic.vet.Vet;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -121,11 +120,14 @@ public class PlayerController {
     }
 
     
-	@DeleteMapping(value = "{playerId}")
+	@DeleteMapping("/{playerId}")
 	@ResponseStatus(HttpStatus.OK)
-	public ResponseEntity<MessageResponse> delete(@PathVariable("playerId") int id) {
-		RestPreconditions.checkNotNull(playerService.getPlayerByUserId(id), "Player", "ID", id);
-		playerService.deletePlayer(id);
-		return new ResponseEntity<>(new MessageResponse("Player deleted!"), HttpStatus.OK);
+	public ResponseEntity<MessageResponse> delete(@PathVariable("playerId") Integer id) {
+		if (!playerService.getPlayerById(id).isPresent())
+            throw new ResourceNotFoundException("Player", "ID", id);
+        else{
+            playerService.deletePlayer(id);
+            return new ResponseEntity<>(new MessageResponse("Player deleted!"), HttpStatus.OK);
+        }
 	}
 }
